Fixed - 	- move input files to donedir Implemented 	- create renamer with default entries, no processing implemented 	- -Overwrite 	- -SkipEncode 	- -MoveDone (Move was previously automatic)

+++ Encode-All.ps1
@@ -7,7 +7,9 @@ param(
     [string]$OutPath = 'D:\VideoProcessing\MKV',
     [string]$LogFileParam = "",
     [bool]$Overwrite = $false,
-    [bool]$SearchNewTracks = $true
+    #[bool]$SearchNewTracks = $true,
+    [bool]$SkipEncode = $false,
+    [bool]$MoveDone = $true
     )
 
     #Specify Handbrake parameters by category
@@ -50,42 +52,56 @@ function Test-FileLock {
 
 Get-ChildItem $InDir\*.iso | ForEach-Object {
     $input=$_
+    $inputBaseName = $_.BaseName
 
-    #Figure out a subdirectory name based on the source
-    $OutDir = $OutPath + '\' + $input.BaseName
+    #Figure out subdirectory names based on the source
+    $OutDir = $OutPath + '\' + $inputBaseName
+
 
     #If a single log wasn't specified, create a per job log file next to the output files
     if ($LogFileParam -match "") {
-        $LogFile = $OutDir+ '\' + $input.BaseName + '.log' 
+        $LogFile = "$OutDir\$inputBaseName.log" 
     }
     else {
         $LogFile = $LogFileParam
     }
 
+    #RenamerFile is a CSV file that holds special instructions about the encode that do not depend on the 
+    #encode destination format.  Used if an entire library gets re-encoded for a new format, without losing renames.
+    #The RenamerFile is kept with the source ISO.  If one is not found, a blank will be created by the encode process.
+    #Examples include renaming numbered tracks to reflect content (ie., Rename "Movie-5.mkv" to "Movie - Cast Interviews.mkv")
+    #The RenamerFile can also contain a Category which, which is a subdirectory the encoded movie will be added to.
+    #For example OutDir = J:\Movies, "Kid Flick" has a category of "Animation", final directory would be
+    #"J:\Movies\Animation\Kid Flick\[encoded files]"
+    $RenamerFile = "$InDir\$inputBaseName.renamer"
+    $RenamerFileExists = Test-Path ($RenamerFile)
     #TempLog is used to hold StdErr during an encode, then is added to the main log file (master log file or per job log file) after the encode is complete.
-    $TempLog = $OutDir+ '\' + $input.BaseName + '.templog'
-        
+    $TempLog = "$OutDir\$inputBaseName.templog"
     #Identify all of the source files so they can be moved after the job.
-    $FileSetFilter = $input.BaseName + ".*"
+    [string]$FileSetFilter = "$inputBaseName.*"
     $FileSet = Get-ChildItem -Path "$InDir" -Filter "$FileSetFilter"
 
     #Don't do anything if the output sub directory already exists.  If you want to redo the encode, just move or delete the directory.
-    if (!(Test-Path $OutDir)){
+    #if (!(Test-Path $OutDir) -and ($SkipEncode){
 
             #Test to see if the input file is locked, such as if it is open in an editor or it is an ISO being ripped directly into $InDir
-            $FileLocked = $false
+            $FileLocked = Test-FileLock ($input)
             if ($FileLocked -notmatch $true){ 
                 #Process a single title at a time.  It tries 100 chapters rather than trying to figure out how many are really there. 
                 for ($a = 1; $a -le 100; $a++){
-        
-                    $output = $OutDir + '\' + $input.BaseName + '-' + ($a -as [string]) +'.mkv'
+                    
+                    if ($a -le 9){$TrackNumberString = "0$a"}
+                        else {$TrackNumberString = $a}
+
+                        
+                    $output = "$OutDir\$inputBaseName-$TrackNumberString.mkv"
                     $FileExists = Test-Path $output
         
                     #Test if file already exists.  Just in case two encoders manage to start in on the same source.
-                    if ($FileExists -eq $false) {
+                    if ($FileExists -match $false -or $Overwrite -match $true) {
 
                         $PathExists = Test-Path $OutDir
-                        if ($PathExists -eq $false){
+                        if ($PathExists -match $false){
 
                             New-Item -ItemType Directory -Force -Path $OutDir
                             $PathExists = Test-Path $OutDir
@@ -93,24 +109,37 @@ Get-ChildItem $InDir\*.iso | ForEach-Object {
                         
                         $cmd_Arguments = "/s /c `"(`"$exe`" $Handbrake_Parameters --input `"$input`" --title $a --output `"$output`")&(echo `"Handbrake $input completion errorlevel = %errorlevel%`">>$TempLog)`""
                         $LogDate = Get-Date; Add-Content $LogFile "`n`n`"$exe`" $Handbrake_Parameters --input `"$input`" --title $a --output `"$output`""
-                    
-                        Start-Process -FilePath c:\windows\system32\cmd.exe -RedirectStandardError $TempLog -Wait -ArgumentList "$cmd_Arguments"
+                        if ($SkipEncode -notmatch $true){
+                            Start-Process -FilePath c:\windows\system32\cmd.exe -RedirectStandardError $TempLog -Wait -ArgumentList "$cmd_Arguments"
+                            Add-Content -Path $LogFile -Value (Get-Content $TempLog); Remove-Item $TempLog
+                            Add-Content $LogFile "`nCMD ErrorLevel $output = $LASTEXITCODE"
+                            }
+                        else {Add-Content $LogFile "`nEncode Skipped by user option.`n"}
 
-
-                        #&$exe $Handbrake_Parameters --input "$_" --title $a --output "$output" 2>> $TempLog
-                        Add-Content -Path $LogFile -Value (Get-Content $TempLog); Remove-Item $TempLog
-                        Add-Content $LogFile "CMD ErrorLevel $output = $LASTEXITCODE"
                         #if ($LASTEXITCODE -eq 1) {$ProcessError = $true}
+                        #Handbrake errors being produced on missing tracks.
                     }
                 }
-
-        #if ($ProcessError -eq $false) {
-        Move-Item $FileSet -Destination $DoneDir -Path $InDir
-
-        #}
+        
+            if ($MoveDone -match $true) {
+                Move-Item "$InDir\$FileSetFilter" -Destination "$DoneDir\"
+            }
 
         }
+    #}
+    #Generate default renamer if one does not exist
+    If ($RenamerFileExists -match $false) {
+    New-Item $RenamerFile -ItemType File
+    Add-Content $RenamerFile "CATEGORY,"
+    Get-ChildItem $OutDir -filter *.mkv | ForEach-Object {
+        $OutputFileBaseName = $_.BaseName 
+        Add-Content $RenamerFile "RENAME,$OutputFileBaseName,$OutputFileBaseName"
+        }
     }
-    #Put Renamer code here
+        
+    if ($MoveDone -match $true) {
+        Move-Item "$InDir\$FileSetFilter" -Destination "$DoneDir\"
+    }
+
 }
